+++ rtfiles/test.rt
@@ -1,4 +1,5 @@
-A  0.3      100,100,200
-L  0,20,10  0.8         255,0,0
-C  0,0,-3    0,0,-1       90
-cy 0,0,-1.1  0,0,-1   4  50 200,200,200
+A   1.              255,255,255
+L   0,0,5  1.0 10,0,255
+C   0,0,0           0,0,1   90
+pl  0,2,5 0.0,-1.0,0.0  100,100,255
+sp  0,0,5           4   222,155,100
